Separated cli and jshint code into their own packages. The main intent of this is to isolate the JSHint code required for those embedding style scenarios where the cli is not required.
All but the npm tests pass, but I’m also unable to get the bin/build script to work - could be an environmental thing here.
