+++ bin/build
@@ -4,7 +4,7 @@
 "use strict";
 
 var browserify = require("browserify");
-var bundle     = browserify("./src/jshint.js");
+var bundle     = browserify("./src/jshint/jshint.js");
 var version    = require("../package.json").version;
 require("shelljs/make");
