+++ bin/jshint
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require("../src/cli.js").interpret(process.argv);
+require("../src/cli/cli.js").interpret(process.argv);
