+++ src/jshint/jshint.js
