+++ src/jshint/reg.js
