+++ src/jshint/state.js
