+++ src/jshint/style.js
