+++ src/jshint/vars.js
