+++ tests/unit/core.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var JSHINT	= require('../../src/jshint.js').JSHINT;
+var JSHINT	= require('../../src/jshint/jshint.js').JSHINT;
 var fs      = require('fs');
 var TestRun = require("../helpers/testhelper").setup.testRun;
