+++ tests/unit/parser.js
@@ -4,7 +4,7 @@
 
 "use strict";
 
-var JSHINT	= require('../../src/jshint.js').JSHINT;
+var JSHINT	= require('../../src/jshint/jshint.js').JSHINT;
 var fs		= require('fs');
 var TestRun = require("../helpers/testhelper").setup.testRun;
 var path    = require("path");
